// csm.web.app/src/app/project/project-common/enums/job-card-status.enum.ts
/*
 * *************************************************************************
 *
 * Copyright: Robert Bosch Power Tools GmbH, 2023
 *
 * *************************************************************************
 */

export enum JobCardStatusEnum {
    Running = 'RUNNING',
    Completed = 'COMPLETED',
    PartlyCompleted = 'PARTLY_COMPLETED',
    Failed = 'FAILED',
}